fix warning C4311: 'type cast': pointer truncation from 'void *' to 'int': all modern compilers have `ptrdiff_t` in their system header files.
TODO: when you must compile with an old MSVC version, you may run into this as 'missing ptrdiff_t type' or words of that order. `typedef size_t ptrdiff_t;` should do it then for pthread-win32 library at least...

+++ pthread_kill.c
@@ -139,7 +139,7 @@ pthread_kill (pthread_t thread, int sig)
               tp->cancelState = PTHREAD_CANCEL_DISABLE;
               ptw32_mcs_lock_release(&stateLock);
 
-              result = TerminateThread(tp->threadH, (DWORD)(int)PTHREAD_CANCELED);
+              result = TerminateThread(tp->threadH, (DWORD)(ptrdiff_t)PTHREAD_CANCELED);
               result = (result != 0) ? 0 : EINVAL;
 
               // Set exit to CANCELED (KILLED) when it hasn't been set already.
